refactor(spf): move findTrack + updateTrackInPresentation to media/utils
Both helpers operate purely on HAM-shaped types from `media/types` and
have no dependency on `core/` (no signals, no effects, no reactors).
They belong in the framework-agnostic media layer per the source-layout
rules in `packages/spf/src/CLAUDE.md`.

New file: `packages/spf/src/media/utils/tracks.ts` with both functions
exported. Naming rationale:

- `media/utils/track-selection.ts` exists for selection-state helpers
  (`SelectedTrackIdKeyByType`, `BufferKeyByType`, `getSelectedTrack`).
  These functions are about track lookup / presentation mutation, not
  selection state — different enough concept to warrant a sibling
  file.
- `tracks.ts` (plural) covers track-within-presentation operations
  generically; future track utilities of similar shape go here.

`resolve-track.ts` updated to import from the new location;
`updateTrackInPresentation` is no longer exported from the behavior
module (it was only ever used internally and now lives in its proper
home).

Tests: resolve-track.test.ts — 7 passed.
Typecheck: clean.
Workspace check: 6/6 passed.
